+++ packages/preview-middleware/src/index.ts
@@ -1,3 +1,4 @@
 export * from './ui5/middleware';
+export { generatePreviewFiles } from './api/config';
 export { FlpSandbox, initAdp } from './base';
 export { FlpConfig, RtaConfig, TestConfig, MiddlewareConfig } from './types';
