+++ packages/preview-middleware/src/ui5/middleware.ts
@@ -3,24 +3,7 @@ import type { RequestHandler } from 'express';
 import type { MiddlewareParameters } from '@ui5/server';
 import { FlpSandbox, initAdp } from '../base/flp';
 import type { MiddlewareConfig } from '../types';
-
-/**
- * The developer mode is only supported for adaptation projects, therefore, notify the user if it is wrongly configured and then disable it.
- *
- * @param config configurations from the ui5.yaml
- * @param logger logger instance
- */
-function sanitizeConfig(config: MiddlewareConfig, logger: ToolsLogger): void {
-    if (config.rta && config.adp === undefined) {
-        config.rta.editors = config.rta.editors.filter((editor) => {
-            if (editor.developerMode) {
-                logger.error('developerMode is ONLY supported for SAP UI5 adaptation projects.');
-                logger.warn(`developerMode for ${editor.path} disabled`);
-            }
-            return !editor.developerMode;
-        });
-    }
-}
+import { sanitizeConfig } from '../api/config';
 
 /**
  * Create the router that is to be exposed as UI5 middleware.
